WIP refactoring the responsibilities of the reference Reasoner.
