+++ test/behaviour/resolution/framework/Reasoner.java
@@ -18,6 +18,8 @@
 
 package com.vaticle.typedb.core.test.behaviour.resolution.framework;
 
+import com.vaticle.typedb.core.common.exception.TypeDBException;
+import com.vaticle.typedb.core.common.iterator.FunctionalIterator;
 import com.vaticle.typedb.core.common.parameters.Arguments;
 import com.vaticle.typedb.core.common.parameters.Context;
 import com.vaticle.typedb.core.common.parameters.Options;
@@ -25,15 +27,19 @@ import com.vaticle.typedb.core.concept.Concept;
 import com.vaticle.typedb.core.concept.answer.ConceptMap;
 import com.vaticle.typedb.core.logic.Rule;
 import com.vaticle.typedb.core.logic.resolvable.Concludable;
+import com.vaticle.typedb.core.pattern.Conjunction;
 import com.vaticle.typedb.core.pattern.Disjunction;
 import com.vaticle.typedb.core.rocks.RocksSession;
 import com.vaticle.typedb.core.rocks.RocksTransaction;
+import com.vaticle.typedb.core.traversal.common.Identifier;
 import com.vaticle.typedb.core.traversal.common.Identifier.Variable;
+import com.vaticle.typeql.lang.query.TypeQLMatch;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 
+import static com.vaticle.typedb.core.common.exception.ErrorMessage.Internal.ILLEGAL_STATE;
 import static com.vaticle.typedb.core.common.iterator.Iterators.iterate;
 import static java.util.Collections.singletonList;
 
@@ -55,8 +61,25 @@ public class Reasoner {
         return reasoner;
     }
 
+    public Map<Conjunction, FunctionalIterator<ConceptMap>> query(TypeQLMatch inferenceQuery) {
+        // TODO: How do we handle disjunctions inside negations?
+        Disjunction disjunction = Disjunction.create(inferenceQuery.conjunction().normalise());
+        tx.logic().typeResolver().resolve(disjunction);
+        HashMap<Conjunction, FunctionalIterator<ConceptMap>> conjunctionAnswers = new HashMap<>();
+        disjunction.conjunctions().forEach(conjunction -> conjunctionAnswers.put(conjunction, traverse(conjunction)));
+        return conjunctionAnswers;
+    }
+
     public Map<String, RuleRecorder> ruleRecorderMap() {
         return ruleRecorders;
+    }
+
+    public RocksTransaction tx() {
+        return tx;
+    }
+
+    public void close() {
+        tx.close();
     }
 
     private void runRules() {
@@ -66,7 +89,7 @@ public class Reasoner {
             reiterateRules = false;
             for (RuleRecorder rule : this.ruleRecorders.values()) {
                 rule.resetRequiresReiteration();
-                runRule(tx, rule);
+                runRule(rule);
                 reiterateRules = reiterateRules || rule.requiresReiteration();
             }
         }
@@ -74,17 +97,22 @@ public class Reasoner {
         // held in memory instead.
     }
 
-    private static void runRule(RocksTransaction tx, RuleRecorder ruleRecorder) {
+    private void runRule(RuleRecorder ruleRecorder) {
         // Get all the places where the rule condition is satisfied and materialise for each
-        tx.reasoner().executeTraversal(
-                new Disjunction(singletonList(ruleRecorder.rule().when())),
-                new Context.Query(tx.context(), new Options.Query()),
-                Utils.filterRetrievableVars(ruleRecorder.rule().when().identifiers())
-        ).forEachRemaining(whenConcepts -> ruleRecorder.rule().conclusion()
+        traverse(ruleRecorder.rule().when())
+                .forEachRemaining(whenConcepts -> ruleRecorder.rule().conclusion()
                 .materialise(whenConcepts, tx.traversal(), tx.concepts())
                 .map(thenConcepts -> new ConceptMap(filterRetrievableVars(thenConcepts)))
                 .filter(ruleRecorder::isInferredAnswer)
                 .forEachRemaining(ans -> ruleRecorder.recordInference(whenConcepts, ans)));
+    }
+
+    private FunctionalIterator<ConceptMap> traverse(Conjunction conjunction) {
+        return tx.reasoner().executeTraversal(
+                new Disjunction(singletonList(conjunction)),
+                new Context.Query(tx.context(), new Options.Query()),
+                filterRetrievableVars(conjunction.identifiers())
+        );
     }
 
     private static Map<Variable.Retrievable, Concept> filterRetrievableVars(Map<Variable, Concept> concepts) {
@@ -96,12 +124,12 @@ public class Reasoner {
         return newMap;
     }
 
-    public void close() {
-        tx.close();
-    }
-
-    public RocksTransaction tx() {
-        return tx;
+    public static Set<Identifier.Variable.Retrievable> filterRetrievableVars(Set<Identifier.Variable> vars) {
+        return iterate(vars).filter(var -> var.isName() || var.isAnonymous()).map(var -> {
+            if (var.isName()) return var.asName();
+            if (var.isAnonymous()) return var.asAnonymous();
+            throw TypeDBException.of(ILLEGAL_STATE);
+        }).toSet();
     }
 
     public static class RuleRecorder {
